Move PC guard to first block
But this doesn't actually work for direct sends... There's not a good
way to do the check we want to do. There's nothing setting the PC to the
actual value we want, so I think the check is actually bogus. We'd have
to eagerly figure out what the PC should be, then check that against
what we compile for.